+++ libs/nrlib/trend/trendkit.hpp
@@ -38,11 +38,11 @@ namespace NRLib {
 #define RMISSING -99999.000
 #define IMISSING -99999
 
-  void EstimateConstantTrend(const std::vector<std::vector<float> >  & blocked_logs,
+  void EstimateConstantTrend(const std::vector<std::vector<double> > & blocked_logs,
                              double                                  & trend);
 
-  void                   PreprocessData0D(const std::vector<std::vector<float> > & blocked_logs,
-                                          std::vector<double>                    & y);
+  void                   PreprocessData0D(const std::vector<std::vector<double> > & blocked_logs,
+                                          std::vector<double>                     & y);
 
   int                    GetTrend1DFileFormat(const std::string & file_name,
                                               std::string       & errText);
@@ -81,7 +81,7 @@ namespace NRLib {
                                          std::string               & errTxt);
 
   void                   PreprocessData1D(const std::vector<std::vector<double> > & s,
-                                          const std::vector<std::vector<float> >  & blocked_logs,
+                                          const std::vector<std::vector<double> > & blocked_logs,
                                           const std::vector<double>               & trend_cube_sampling,
                                           std::vector<double>                     & x,
                                           std::vector<double>                     & y);
@@ -143,7 +143,7 @@ namespace NRLib {
 
   void                   PreprocessData2D(const std::vector<std::vector<double> > & s1,
                                           const std::vector<std::vector<double> > & s2,
-                                          const std::vector<std::vector<float> >  & blocked_logs,
+                                          const std::vector<std::vector<double> > & blocked_logs,
                                           const std::vector<std::vector<double> > & trend_cube_sampling,
                                           std::vector<double>                     & x,
                                           std::vector<double>                     & y,
